# faster decoding speed with Visual