@Cyan4973 2019-12-03
CI found a completely unrelated bug,
in "LZ4_compress_HC_continue with an external dictionary",
that now requires further investigation.