@ffulin 2020-04-04
@Cyan4973 - I made this same change locally against 1.9.2, and while preparing the PR I found this was already fixed in head by this change :). I thought I'd share my timings in case they are interesting to you.

I saw:
 - some cases improved across the board at all compression levels (like the numbers reported here)
 - some cases not improved at all
 - some cases, when using high acceleration values, improved by almost 2x!

(all numbers are x64 MSVC 2019 v16.5.1 on Ryzen 3900X. Each test repeated 2000 times)

![image](https://user-images.githubusercontent.com/11589403/78462085-9f157b80-7683-11ea-8776-4bede859e697.png)